Code runs and works!
Note: this was done ~contemporaneously with the previous commits. Just
noticed I hadn't checked it in.

I manually wrote out the code in machine code (not really familiar with
ARM, was jumping around the various manuals). Failed attempt is
commented out. The code that runs on the LPC810 just configures the UART
and keeps writing a single char to the the UART output.

It's really annoying to manually relocate code.